refactor: extract numbers content to component
